(Major) Changed the time type to float from int. Now time is in seconds, instead of milliseconds as before. Allows for better accuracy and compatibility with how other subsystems use time

+++ source/poro/default_application.h
@@ -8,34 +8,34 @@
 //=============================================================================
 namespace poro {
 
-class DefaultApplication : 
-	public IApplication, 
-	public IMouseListener, 
+class DefaultApplication :
+	public IApplication,
+	public IMouseListener,
 	public IKeyboardListener
 {
 public:
 	DefaultApplication() : mInitializedListeners( false ) { }
 	virtual ~DefaultApplication() { }
 
-	//-------------------------------------------------------------------------	
-	
-	virtual void Update( int deltaTime ) { }
+	//-------------------------------------------------------------------------
+
+	virtual void Update( float deltaTime ) { }
 	virtual void Draw( poro::IGraphics* graphics ) { }
 
 	virtual void	Init();
 	virtual void	Exit();
 
-	//-------------------------------------------------------------------------	
-	
+	//-------------------------------------------------------------------------
+
 	virtual void MouseMove(const poro::types::vec2& pos) { }
 	virtual void MouseButtonDown(const poro::types::vec2& pos, int button) { }
 	virtual void MouseButtonUp(const poro::types::vec2& pos, int button) { }
-	
+
 	virtual void OnKeyDown( int key, poro::types::charset unicode ) { }
 	virtual void OnKeyUp( int key, poro::types::charset unicode ) { }
 
-	//-------------------------------------------------------------------------	
-	
+	//-------------------------------------------------------------------------
+
 private:
 	bool mInitializedListeners;
 };
